requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0

# PDF processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pdfminer.six>=20221105
pdf2image>=1.16.0

# OCR and image processing
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Table extraction
camelot-py>=0.11.0
tabula-py>=2.8.0

# Date and text processing
python-dateutil>=2.8.0
regex>=2023.0.0

# Web interface
streamlit>=1.28.0

# Utilities
python-magic>=0.4.0
chardet>=5.0.0

# Development and testing
pytest>=7.0.0
black>=23.0.0 